Configure application on root path
It can be useful to configure your SPA on the root path. Especailly when
using lazy loaded modules in angular. Previously that was not possible,
as the empty path caused some problems with the implementation.

Now the root path is dedected, and configured properly. If no other path
matches, the root path will checked for matching resources, if any.