Add whitelist to AssetController
Previously the AssetController served all files that where requested
without checking if they are allowed to be servced. In fact, one could
reqeust every file in the classpath and ServletContext. Meaning an
attacker could download class or configuration files stored in the
classpath.

This commit adds a whitelist of folders, that can be used to serve
resources. If a folder is not in this list, a HTTP 404 will be returned
to the requester.

(cherry picked from commit bbc05f0232a6b39f0959e4dfd78d380365b84832)